Adjusted Attributes Order For All Entities
Revamped attribute order so as to provide for more important and relevant information to appear at or near the top of each search result. Object name, description, and aliases are now prioritized for all entities. Did this by setting EntityTypeRenderer to include all entities and added f_common.topic to the type_map Hashmap and getBestType method in the EntityTypeRenderer class.